Game Back in Order + Pokemon Class Started
The game now works as completely as intended, for now. Meaning it goes
from Intro to Main Menu to New Game and back to Intro, loading,
appending, or creating files as needed. A brief test of the Pokemon
class showed that data is being correctly saved. However, the most
recent expansion has not be tested.