Summary Screen Info Started + Moved Input Gathering
Info (first screen) of summary screen populates with correct
information, except types, item, and remaining xp because no functions
exist for those yet. All input for the team screen was moved to a
separate function to reduce cluttering. All input will be moved into
that function for easier updating in the future.