Began Adding Ribbon Functionality
Ribbons now load and display properly, and are accessible under the
pokemon submenu, as opposed to the usual summary region. To have added
it there, I would have had to redo the graphics and that did not appeal
to me, especially after I got it all to work. Next thing to add is a
description for the ribbons.