Fixed Mouse Clicking Bugs and Features
SetSummaryPage and SetGameState now run at the end of the frame, to
avoid conflicts with GatherInput looking for the same input
(GetMouseUp). Added mouse wheel functionality for ribbons and summary
screen to rotate through the team roster. Next update will move all RNG
to a SystemManager function that will enable it to reseed if the same
number comes up.