Updated to Unity 5.3.4f1
Updated the project in order to get rid of the input caret not vanishing
problem. This resulted in the depreciation of some functions, namely
Application.LoadLevel, Application.loadedLevelName, and File.Replace. I
have corrected the first two problems, and will need to figure out
what's going on with the last.