Patch System + Added Status to Debug
Fixed up the patch system so that each new version will initialize new
fields,  and create a backup before patching the new file. Barring
dramatic changes, this should allow files to be upgradable and
backwards-compatable through Version Tolerant Serialization. Status can
now be given to a pokemon through the debug options, and they are
correctly displayed in the team and in the summary.