Fixed Berries Omission + Continued Shop
Finished purchase interface to interact with keyboard and mouse as well
as added and corrected functionality for all purchase interface buttons,
the item sprite, item name, and total cost. Filters now work and the
next and previous buttons account for when the list of items to display
is a multiple of 15, which prevents it from loading an entire blank
page.