Added Bag to Battle Scene
Clicking bag will now display the bag and it's contents. The arrow keys,
mouse wheel, and mouse buttons all respond appropriately. Fixed an issue
in Main Game where clicking a bag icon would cause the player to select
the first item in the chosen slot. This was due to a delay in the pocket
change animation. To remedy this, I added a new boolean to determine if
a pocket change is in progress. If it is, the ProcessGiveitem function
fails.