Implemented Summary Screen and Ribbon Screen
The summary screen and ribbon screen are viewable from battle. In
addition, the field effect was reworked so any boosts separate from the
field effect will be retained, and only those boosts from the field will
be altered depending on who gets switched in. Any boosts that the
opponent had prior to switching are also reverted back to zero and
recalculated.