Began Damage Calculation
Resolving an attack currently decreases PP,  checks for a protection
effect, checks for a negating ability, calculates the type modifier of
the attack, calculates the accuracy of the attack, and calculates if the
attack is a critical. Actual damage and damage boosters (items, ability,
field) as well as recoil and a faint check, still need to be
implemented.