Fixed TypeMod Error + Switch Error
Turns out the damage issue was caused due to faulty handling of a
resisted modifier. The switch error was caused from the state not being
switched once a Pokemon was chosen. I also called the FIllInBattlerData
function to show the new pokemon as soon as it's in. This will be
changed to an animation once it is made.