Added Animations
There is now a faint and a send out animation for each player. The code
still needs to be updated so that the displayed team will reflect the
newly fainted pokemon. Also, an end battle function still needs to be
made so that the appropriate points, badge, and items are given.