Added EndFight function
Once a combatant has no remaining unfainted pokemon, the game calls the
EndFight function. This processes the speech, the points prize, and any
other items or badges prizes, plus increments the victory count. More
work needs to be done to finish it, however the fact that it works is
important.