Finished BattleScene
The Battle Scene works as expected. There are a few bugs that I can't
figure out (for instance, animations for sending out Pokemon end before
the material is changed back to normal),. but otherwise the battle flow
works correctly, animations play at the appropriate time, and prizes are
given and regulated.